Medication Name into Content Title
For now, in order to get the reminder ID from med name, I will have to place the medication Name in the Content Title.
I would place it in the setContentText method and then call getContentText but the latter method has been deprecated. I also added the Ignore and Acknowledge buttons